LayoutTests/fast/canvas/toDataURL-unsupportedTypes.html

<!DOCTYPE html>
<head>
    <script src="../../resources/js-test-pre.js"></script>
</head>
<body>
    <canvas width="8" height="8"></canvas>
    <script>
        var canvas = document.querySelector("canvas");
        canvas.width = 8;
        canvas.height = 8;

        description("This is a test of the unsupported mime-types for canvas.toDataURL().");
        shouldBe("canvas.toDataURL('image/heic')", "canvas.toDataURL('image/png')");
        shouldBe("canvas.toDataURL('image/heic-sequence')", "canvas.toDataURL('image/png')");
        shouldBe("canvas.toDataURL('image/vnd.adobe.photoshop')", "canvas.toDataURL('image/png')");
        shouldBe("canvas.toDataURL('application/pdf')", "canvas.toDataURL('image/png')");
        shouldBe("canvas.toDataURL('image/targa')", "canvas.toDataURL('image/png')");
    </script>
    <script src="../../resources/js-test-post.js"></script>
</body>
