LayoutTests/fast/canvas/toDataURL-unsupportedTypes-expected.txt

This is a test of the unsupported mime-types for canvas.toDataURL().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS canvas.toDataURL('image/heic') is canvas.toDataURL('image/png')
PASS canvas.toDataURL('image/heic-sequence') is canvas.toDataURL('image/png')
PASS canvas.toDataURL('image/vnd.adobe.photoshop') is canvas.toDataURL('image/png')
PASS canvas.toDataURL('application/pdf') is canvas.toDataURL('image/png')
PASS canvas.toDataURL('image/targa') is canvas.toDataURL('image/png')
PASS successfullyParsed is true

TEST COMPLETE
